move grid component to the axis
