+++ packages/dx-react-chart-bootstrap4/src/index.js
@@ -11,8 +11,6 @@ export { PieSeries } from './plugins/pie-series';
 export { ScatterSeries } from './plugins/scatter-series';
 export { SplineSeries } from './plugins/spline-series';
 
-export { ValueGrid } from './plugins/value-grid';
-export { ArgumentGrid } from './plugins/argument-grid';
 export { ValueAxis } from './plugins/value-axis';
 export { ArgumentAxis } from './plugins/argument-axis';
