+++ packages/dx-react-chart-bootstrap4/src/plugins/argument-axis.jsx
@@ -4,7 +4,8 @@ import { Root } from '../templates/axis/root';
 import { Tick } from '../templates/axis/tick';
 import { Label } from '../templates/axis/label';
 import { Line } from '../templates/axis/line';
+import { Grid } from '../templates/axis/grid';
 
 export const ArgumentAxis = withComponents({
-  Root, Tick, Label, Line,
+  Root, Tick, Label, Line, Grid,
 })(ArgumentAxisBase);
