--- packages/dx-react-chart-bootstrap4/src/plugins/argument-grid.jsx
@@ -1,5 +1,0 @@
-import { withComponents } from '@devexpress/dx-react-core';
-import { ArgumentGrid as ArgumentGridBase } from '@devexpress/dx-react-chart';
-import { Line } from '../templates/grid/line';
-
-export const ArgumentGrid = withComponents({ Line })(ArgumentGridBase);
