+++ packages/dx-react-chart-bootstrap4/src/templates/axis/grid.jsx
@@ -1,6 +1,6 @@
-import { Grid } from '@devexpress/dx-react-chart';
+import { Axis } from '@devexpress/dx-react-chart';
 import { withClassName } from '../utils';
 
-export const Line = withClassName(
+export const Grid = withClassName(
   'dx-c-bs4-stroke-current-color', 'dx-c-bs4-crisp-edges', 'dx-c-bs4-axis-opacity',
-)(Grid.Line);
+)(Axis.Grid);
