+++ packages/dx-react-chart-material-ui/src/index.js
@@ -9,8 +9,6 @@ export { AreaSeries } from './plugins/area-series';
 export { SplineSeries } from './plugins/spline-series';
 export { ScatterSeries } from './plugins/scatter-series';
 
-export { ValueGrid } from './plugins/value-grid';
-export { ArgumentGrid } from './plugins/argument-grid';
 export { ValueAxis } from './plugins/value-axis';
 export { ArgumentAxis } from './plugins/argument-axis';
