+++ packages/dx-react-chart/src/index.js
@@ -15,7 +15,6 @@ export { ScatterSeries } from './plugins/scatter-series';
 export { SplineSeries } from './plugins/spline-series';
 
 export { Axis, ArgumentAxis, ValueAxis } from './plugins/axis';
-export { Grid, ArgumentGrid, ValueGrid } from './plugins/grid';
 export { Tooltip } from './plugins/tooltip';
 
 export * from './plugins/event-tracker';
