+++ packages/dx-react-chart/src/templates/axis/grid.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
 
-export class Line extends React.PureComponent {
+export class Grid extends React.PureComponent {
   render() {
     const {
       x1, x2, y1, y2, ...restProps
@@ -15,7 +15,7 @@ export class Line extends React.PureComponent {
   }
 }
 
-Line.propTypes = {
+Grid.propTypes = {
   x1: PropTypes.number.isRequired,
   x2: PropTypes.number.isRequired,
   y1: PropTypes.number.isRequired,
